Add resource to log format options
Cleaning up log messages and thought it
would be useful to make it "easier" to log the
volume/snapshot/backup/image id's more easily.

Unlike the Nova Instance work that was submitted this
key will work with any db ref object that has a "name"
member.  Most OpenStack objects have an internal name
that is built using the UUID of the resource object.

Resource objects in Cinder for example use:
  "<resourceName-uuid>"

So it seems like a pretty natural fit for this. Most
other projects have the same model (including Nova,
although Nova specific code was added for this already).

With this change, LOG calls in projects can be sent like:
  LOG.info(_LI('begin some operation'), resource=dbref)

The result will be the resource "Name" prepended to the msg:
  INFO cinder.volume.manager [-]\
  [volume-40e0f194-0ccb-422f-a543-c5f9022acb6f] \
  succesfully deleted volume

Change-Id: I7745f6632147248b5180978b51119014783209bd
Partially-Implements: blueprint https://blueprints.launchpad.net/oslo.log/+spec/app-agnostic-logging-parameters