Added graceful argument on Service.stop method
This change adds a graceful shutdown arugment to the Service.stop
method in order to allow a service to gracefully wait for threads to
shutdown. Since threadgroup already supports a graceful shutdown
argument for its stop() method, it is trivial to support.

Added unit tests to demonstrate the expected behavior between graceful
and ungraceful service stop()s.

Change-Id: Ibc0df09daef5e2a07f232543f44f2fc4d203511c
Closes-Bug: 1406638