Log missing dir message only once
Some services that use the policy code create new Enforcers
frequently.  This means that if the policy_dirs are not found a
log message is emitted each time, which results in an excessive
amount of noise in the logs.  Instead, we only want to log the
missing dirs warning once per missing path.

Change-Id: I6d4dcdf7e61d8bbb7084068c7288077cb2e1d193
Closes-Bug: 1421863