Fix the regex for turning project.lib back to oslo.lib
The existing pattern was using '.' to catch any character, which
included spaces and underscore in some projects although we were
assuming it would only be import statements. This patch changes the
pattern to remember what the character is and use it in the replacement
pattern to avoid breaking symbol names that have "oslo" in them or other
similar cases.

Closes-Bug: #1423361
Change-Id: I0f6995c4204ff8690f17ee97b9b7f9c2042e0cf6