Add dates to unreleased change report
Add an option to show date of the commit to the release notes so the
reporting tool for showing the unreleased changes can use it to indicate
how old different commits are.

Also replace the path to the repository with the name of the project in
the release notes header for each project.

Change-Id: I115366590d39ffaee3c3fc39a71b87c3f03d0837