Clean up logging to conform to guidelines
Correct instances where guidelines are not being followed.

Guidelines can be found here:
http://docs.openstack.org/developer/oslo.i18n/guidelines.html

Some of this is seen by folks implementing new code and results
in confusion during reviews. These are mostly trivial changes,
but the hope is this will provide good examples so as to avoid
that confusion.

Also cleaning up case of deprecated LOG.warn called in leiu of
LOG.warning as well as redundant cases of formatting an exception
inside of LOG.exception.

Change-Id: I7decb7bf5f3185f7d26ad99b0cb3475a88d3ec99