Just use os.urandom instead of Crypto.Random
Crypto.Random needs someone to keep track of os.fork() and
call Crypto.Random.atfork() in both the parent and the child
Since this is too onerous a burden on the loosely coupled
oslo code and the consuming project. It's just simpler to switch
to os.urandom() to get the bytes we need.

Closes-Bug: #1464037
Change-Id: I09d7d2708dafdb3ac2514844e74445d93037412a