Fix bad mock assertion calis
Latest mock checks for assertion calls that don't exist, usually because
they are typos in the tests and with the nature of mock before it just
lets you create whatever if you're not using a spec for the mock.

Fixing the busted assertions exposed bugs in the tests so those are also
fixed.

Closes-Bug: #1482377

Change-Id: I37a546d9f8679f10d50f94774280ff02195a29f9