Memory cache cleanup thread
Instead of cleaning up time-based objects in the memory cache,
provide a thread that wakes up periodically to clean the cache.
This thread only will be started if objects present in the cache
have a expiration value.

By default, the thread will wake up on a 60-second interval to
loop through the cache list and cleanup any stale items.  This
allows the get() method from having to traverse the list every
single time it's called, thus improving the performance on the
retrieval from the cache.

Change-Id: I97dcd55838941de4bf31aa84b4af7b2cb8e4f116
Closes-Bug: 1483322