Fix the bug of can't get the desired image info
The current imageutils.QemuImgInfo class parse all the image info
at first, but if one of the info is unavailable, it would raise a
ValueError.

Sometimes, user only care one of the info which is really
available, but it may be failed because of the ValueError which is
caused by another unavailable info.

This patch resolves the problem by parsing each of the info separately,
so that getting the desired image info will not be influenced by other
unavailable info.

I have committed this patch to cinder:
Change-Id: Ic5f3b80e172baaafc07432319f8aa24bf117b392

Change-Id: I3692da15db6ee99edceb856643628b76a28ad270
Closes-Bug: #1516206